requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
mysql-connector-python==8.0.30
Werkzeug==2.2.2 